Fix caption length issue
NetSuite limits the caption length to be <= 40 chars. It also defaults this field on file creation through web services to be the filename. So this fix will edit the file after creation if the filename is > 40 chars and truncate down the caption to be the first 40 chars of the filename. Without this, the initial creation of the file will work, but subsequent upload attempts will fail saying the caption cannot be > than 40 chars.